+++ tests/examples/query_param_tests.rs
@@ -7,7 +7,6 @@ use httpmock::MockServer;
 #[test]
 fn url_matching_test() {
     // Arrange
-    let _ = env_logger::try_init();
     let server = MockServer::start();
 
     let m = server.mock(|when, then| {